vdk-structlog: move the log level util functions to separate file
